% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{add}
\alias{add}
\title{Forecast}
\usage{
add(x, y)
}
\arguments{
\item{x, y}{Values to add together}
}
\value{
The \code{numeric} sum of \code{x} and \code{y}
}
\description{
Adds \code{x} and \code{y} together.
}
